% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{AbstractAuthentication}
\alias{AbstractAuthentication}
\title{AbstractAuthentication}
\description{
Base R6 class for authentication methods
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{auth_token}}{The authentication token}

\item{\code{config}}{The client configuration}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AbstractAuthentication-new}{\code{AbstractAuthentication$new()}}
\item \href{#method-AbstractAuthentication-has_expired}{\code{AbstractAuthentication$has_expired()}}
\item \href{#method-AbstractAuthentication-refresh}{\code{AbstractAuthentication$refresh()}}
\item \href{#method-AbstractAuthentication-get_headers}{\code{AbstractAuthentication$get_headers()}}
\item \href{#method-AbstractAuthentication-clone}{\code{AbstractAuthentication$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractAuthentication-new"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractAuthentication-new}{}}}
\subsection{Method \code{new()}}{
Create a new AbstractAuthentication object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractAuthentication$new(config)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config}}{The client configuration}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new AbstractAuthentication object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractAuthentication-has_expired"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractAuthentication-has_expired}{}}}
\subsection{Method \code{has_expired()}}{
Check if the token has expired
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractAuthentication$has_expired()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TRUE if expired, FALSE otherwise
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractAuthentication-refresh"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractAuthentication-refresh}{}}}
\subsection{Method \code{refresh()}}{
Refresh the authentication token
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractAuthentication$refresh()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractAuthentication-get_headers"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractAuthentication-get_headers}{}}}
\subsection{Method \code{get_headers()}}{
Get the headers for HTTP requests
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractAuthentication$get_headers()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A named character vector of headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AbstractAuthentication-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AbstractAuthentication-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractAuthentication$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
