% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-auth_token.R
\name{auth_token_exists}
\alias{auth_token_exists}
\title{Check if saved token exists}
\usage{
auth_token_exists(config, filename)
}
\arguments{
\item{config}{The client config}

\item{filename}{The filename to check}
}
\value{
TRUE if the file exists, FALSE otherwise
}
\description{
Check if a token file exists
}
