% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{auth_using}
\alias{auth_using}
\title{Authenticate using an authenticator}
\usage{
auth_using(authenticator, config = NULL, name = NULL)
}
\arguments{
\item{authenticator}{The authentication object}

\item{config}{The client configuration (optional)}

\item{name}{The client name (optional)}
}
\value{
An EbxClient object
}
\description{
Low-level authentication function
}
