% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{auth_using_creds}
\alias{auth_using_creds}
\title{Authenticate using saved credentials}
\usage{
auth_using_creds(filename = API_SECRETS_FILE, name = NULL, config = NULL)
}
\arguments{
\item{filename}{The credentials filename (optional)}

\item{name}{The client name (optional)}

\item{config}{The client configuration (optional)}
}
\value{
An EbxClient object
}
\description{
Authenticate using saved OAuth credentials from disk
}
