% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{auth_using_env}
\alias{auth_using_env}
\title{Authenticate using environment variable}
\usage{
auth_using_env(name = NULL, config = NULL)
}
\arguments{
\item{name}{The client name (optional)}

\item{config}{The client configuration (optional)}
}
\value{
An EbxClient object
}
\description{
Authenticate using the EBX_API_TOKEN environment variable
}
