% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{auth_using_oauth}
\alias{auth_using_oauth}
\title{Authenticate using OAuth}
\usage{
auth_using_oauth(
  client_id = NULL,
  client_secret = NULL,
  name = NULL,
  config = NULL
)
}
\arguments{
\item{client_id}{The client ID (optional, can be from env)}

\item{client_secret}{The client secret (optional, can be from env)}

\item{name}{The client name (optional)}

\item{config}{The client configuration (optional)}
}
\value{
An EbxClient object
}
\description{
Authenticate using OAuth client credentials
}
