% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-auth_token.R
\name{AuthToken}
\alias{AuthToken}
\title{AuthToken}
\description{
R6 class for authentication tokens
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{token}}{The access token}

\item{\code{expires}}{The expiration datetime}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AuthToken-new}{\code{AuthToken$new()}}
\item \href{#method-AuthToken-save}{\code{AuthToken$save()}}
\item \href{#method-AuthToken-to_list}{\code{AuthToken$to_list()}}
\item \href{#method-AuthToken-to_json}{\code{AuthToken$to_json()}}
\item \href{#method-AuthToken-clone}{\code{AuthToken$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthToken-new"></a>}}
\if{latex}{\out{\hypertarget{method-AuthToken-new}{}}}
\subsection{Method \code{new()}}{
Create a new AuthToken object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthToken$new(token = NULL, expires = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{token}}{The access token}

\item{\code{expires}}{The expiration datetime (as POSIXct or string)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new AuthToken object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthToken-save"></a>}}
\if{latex}{\out{\hypertarget{method-AuthToken-save}{}}}
\subsection{Method \code{save()}}{
Save the token to disk
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthToken$save(config, filename)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config}}{The client config}

\item{\code{filename}}{The filename to save to}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
self (invisible)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthToken-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-AuthToken-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert AuthToken to a plain list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthToken$to_list(include_token = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{include_token}}{Whether to include the actual token (default: FALSE for security)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list containing auth token fields
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthToken-to_json"></a>}}
\if{latex}{\out{\hypertarget{method-AuthToken-to_json}{}}}
\subsection{Method \code{to_json()}}{
Convert AuthToken to JSON string
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthToken$to_json(pretty = TRUE, include_token = FALSE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{Whether to pretty-print the JSON (default: TRUE)}

\item{\code{include_token}}{Whether to include the actual token (default: FALSE for security)}

\item{\code{...}}{Additional arguments passed to jsonlite::toJSON}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A JSON string representation of the auth token
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthToken-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AuthToken-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthToken$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
