% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{BasicAuth}
\alias{BasicAuth}
\title{BasicAuth}
\description{
Authentication using username and password
}
\section{Super class}{
\code{ebx::AbstractAuthentication} -> \code{BasicAuth}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{email}}{The email address}

\item{\code{password}}{The password}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BasicAuth-new}{\code{BasicAuth$new()}}
\item \href{#method-BasicAuth-has_expired}{\code{BasicAuth$has_expired()}}
\item \href{#method-BasicAuth-refresh}{\code{BasicAuth$refresh()}}
\item \href{#method-BasicAuth-get_headers}{\code{BasicAuth$get_headers()}}
\item \href{#method-BasicAuth-clone}{\code{BasicAuth$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasicAuth-new"></a>}}
\if{latex}{\out{\hypertarget{method-BasicAuth-new}{}}}
\subsection{Method \code{new()}}{
Create a new BasicAuth object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicAuth$new(config, email, password)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config}}{The client configuration}

\item{\code{email}}{The email address}

\item{\code{password}}{The password}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new BasicAuth object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasicAuth-has_expired"></a>}}
\if{latex}{\out{\hypertarget{method-BasicAuth-has_expired}{}}}
\subsection{Method \code{has_expired()}}{
Check if token has expired (always FALSE for basic auth)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicAuth$has_expired()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
FALSE
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasicAuth-refresh"></a>}}
\if{latex}{\out{\hypertarget{method-BasicAuth-refresh}{}}}
\subsection{Method \code{refresh()}}{
Refresh (not supported for basic auth)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicAuth$refresh()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasicAuth-get_headers"></a>}}
\if{latex}{\out{\hypertarget{method-BasicAuth-get_headers}{}}}
\subsection{Method \code{get_headers()}}{
Get the headers for HTTP requests
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicAuth$get_headers()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A named character vector of headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BasicAuth-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BasicAuth-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BasicAuth$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
