% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{ClientConfig}
\alias{ClientConfig}
\title{ClientConfig}
\description{
R6 class for Earth Blox API client configuration
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{base_url}}{The base URL for the API}

\item{\code{api_prefix}}{The version prefix for the API}

\item{\code{oauth_path}}{The path for the OAuth flow}

\item{\code{persistence_driver}}{The persistence driver for storing credentials}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ClientConfig-new}{\code{ClientConfig$new()}}
\item \href{#method-ClientConfig-get_api_base_url}{\code{ClientConfig$get_api_base_url()}}
\item \href{#method-ClientConfig-get_oauth_url}{\code{ClientConfig$get_oauth_url()}}
\item \href{#method-ClientConfig-get_persistence_driver}{\code{ClientConfig$get_persistence_driver()}}
\item \href{#method-ClientConfig-clone}{\code{ClientConfig$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClientConfig-new"></a>}}
\if{latex}{\out{\hypertarget{method-ClientConfig-new}{}}}
\subsection{Method \code{new()}}{
Create a new ClientConfig object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClientConfig$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new ClientConfig object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClientConfig-get_api_base_url"></a>}}
\if{latex}{\out{\hypertarget{method-ClientConfig-get_api_base_url}{}}}
\subsection{Method \code{get_api_base_url()}}{
Get the full API base URL including version prefix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClientConfig$get_api_base_url()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The full API base URL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClientConfig-get_oauth_url"></a>}}
\if{latex}{\out{\hypertarget{method-ClientConfig-get_oauth_url}{}}}
\subsection{Method \code{get_oauth_url()}}{
Get the OAuth URL
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClientConfig$get_oauth_url()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The OAuth URL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClientConfig-get_persistence_driver"></a>}}
\if{latex}{\out{\hypertarget{method-ClientConfig-get_persistence_driver}{}}}
\subsection{Method \code{get_persistence_driver()}}{
Get the persistence driver
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClientConfig$get_persistence_driver()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The persistence driver
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClientConfig-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ClientConfig-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClientConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
