% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-services.R
\name{create_oauth_client}
\alias{create_oauth_client}
\title{Create OAuth Client}
\usage{
create_oauth_client(email, password, name, description = "", scopes = c())
}
\arguments{
\item{email}{The email address of the user}

\item{password}{The password of the user}

\item{name}{The name for the new client}

\item{description}{The description for the new client (optional)}

\item{scopes}{The scopes for the new client (optional)}
}
\value{
An OAuthClient object
}
\description{
Register a new OAuth client with the API
}
