% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-runs.R
\name{get_charts}
\alias{get_charts}
\title{Get Charts}
\usage{
get_charts(run_id, filter = NULL)
}
\arguments{
\item{run_id}{The run ID}

\item{filter}{Optional filter to apply to chart titles}
}
\value{
A list of chart data
}
\description{
Get charts from a run
}
