% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{get_client}
\alias{get_client}
\title{Get or create a client}
\usage{
get_client(name = NULL)
}
\arguments{
\item{name}{The client name (optional)}
}
\value{
An EbxClient object
}
\description{
Get an existing client or the current default client
}
