% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-runs.R
\name{get_layers}
\alias{get_layers}
\title{Get Layers}
\usage{
get_layers(run_id, filter = NULL)
}
\arguments{
\item{run_id}{The run ID}

\item{filter}{Optional filter to apply to layer titles}
}
\value{
A list of layer data
}
\description{
Get layers from a run
}
