% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-projects.R
\name{get_project}
\alias{get_project}
\title{Get Project}
\usage{
get_project(project_id)
}
\arguments{
\item{project_id}{The project ID}
}
\value{
A Project object
}
\description{
Get a specific project by ID
}
