% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-runs.R
\name{get_run_status}
\alias{get_run_status}
\title{Get Run Status}
\usage{
get_run_status(run_id)
}
\arguments{
\item{run_id}{The run ID}
}
\value{
The run status as a string
}
\description{
Get the status of a specific run
}
