% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-runs.R
\name{get_tables}
\alias{get_tables}
\title{Get Tables}
\usage{
get_tables(run_id, filter = NULL)
}
\arguments{
\item{run_id}{The run ID}

\item{filter}{Optional filter to apply to table titles}
}
\value{
A list of table data
}
\description{
Get tables from a run
}
