% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-auth_token.R
\name{load_auth_token}
\alias{load_auth_token}
\title{Load Auth Token from disk}
\usage{
load_auth_token(config, filename)
}
\arguments{
\item{config}{The client config}

\item{filename}{The filename to load from}
}
\value{
An AuthToken object
}
\description{
Load an auth token from a saved file
}
