% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-oauth_client.R
\name{load_oauth_client}
\alias{load_oauth_client}
\title{Load OAuth Client from disk}
\usage{
load_oauth_client(config = NULL, filename = API_SECRETS_FILE)
}
\arguments{
\item{config}{The client config (optional)}

\item{filename}{The filename to load from (optional)}
}
\value{
An OAuthClient object
}
\description{
Load an OAuth client from a saved file
}
