% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence.R
\name{LocalFilePersistence}
\alias{LocalFilePersistence}
\title{LocalFilePersistence}
\description{
R6 class for persisting credentials to local filesystem
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{The directory path for storing files}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LocalFilePersistence-new}{\code{LocalFilePersistence$new()}}
\item \href{#method-LocalFilePersistence-save}{\code{LocalFilePersistence$save()}}
\item \href{#method-LocalFilePersistence-load}{\code{LocalFilePersistence$load()}}
\item \href{#method-LocalFilePersistence-exists}{\code{LocalFilePersistence$exists()}}
\item \href{#method-LocalFilePersistence-clone}{\code{LocalFilePersistence$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LocalFilePersistence-new"></a>}}
\if{latex}{\out{\hypertarget{method-LocalFilePersistence-new}{}}}
\subsection{Method \code{new()}}{
Create a new LocalFilePersistence object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalFilePersistence$new(path = API_SECRETS_PATH)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{The directory path for storing files}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new LocalFilePersistence object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LocalFilePersistence-save"></a>}}
\if{latex}{\out{\hypertarget{method-LocalFilePersistence-save}{}}}
\subsection{Method \code{save()}}{
Save data to a file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalFilePersistence$save(filename, data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{The name of the file}

\item{\code{data}}{The data to save (will be converted to JSON)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL (invisible)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LocalFilePersistence-load"></a>}}
\if{latex}{\out{\hypertarget{method-LocalFilePersistence-load}{}}}
\subsection{Method \code{load()}}{
Load data from a file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalFilePersistence$load(filename)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{The name of the file}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The loaded data as a list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LocalFilePersistence-exists"></a>}}
\if{latex}{\out{\hypertarget{method-LocalFilePersistence-exists}{}}}
\subsection{Method \code{exists()}}{
Check if a file exists
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalFilePersistence$exists(filename)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{The name of the file}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
TRUE if the file exists, FALSE otherwise
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LocalFilePersistence-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LocalFilePersistence-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalFilePersistence$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
