% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-oauth_client.R
\name{oauth_credentials_exist}
\alias{oauth_credentials_exist}
\title{Check if saved OAuth credentials exist}
\usage{
oauth_credentials_exist(config = NULL, filename = API_SECRETS_FILE)
}
\arguments{
\item{config}{The client config (optional)}

\item{filename}{The filename to check (optional)}
}
\value{
TRUE if the file exists, FALSE otherwise
}
\description{
Check if a credentials file exists
}
