% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{OAuthAuthentication}
\alias{OAuthAuthentication}
\title{OAuthAuthentication}
\description{
Authentication using OAuth client credentials
}
\section{Super class}{
\code{ebx::AbstractAuthentication} -> \code{OAuthAuthentication}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{client_id}}{The client ID}

\item{\code{client_secret}}{The client secret}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OAuthAuthentication-new}{\code{OAuthAuthentication$new()}}
\item \href{#method-OAuthAuthentication-get_token_filename}{\code{OAuthAuthentication$get_token_filename()}}
\item \href{#method-OAuthAuthentication-load_saved_credentials}{\code{OAuthAuthentication$load_saved_credentials()}}
\item \href{#method-OAuthAuthentication-save_credentials}{\code{OAuthAuthentication$save_credentials()}}
\item \href{#method-OAuthAuthentication-refresh}{\code{OAuthAuthentication$refresh()}}
\item \href{#method-OAuthAuthentication-get_headers}{\code{OAuthAuthentication$get_headers()}}
\item \href{#method-OAuthAuthentication-clone}{\code{OAuthAuthentication$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ebx" data-topic="AbstractAuthentication" data-id="has_expired"><a href='../../ebx/html/AbstractAuthentication.html#method-AbstractAuthentication-has_expired'><code>ebx::AbstractAuthentication$has_expired()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OAuthAuthentication-new"></a>}}
\if{latex}{\out{\hypertarget{method-OAuthAuthentication-new}{}}}
\subsection{Method \code{new()}}{
Create a new OAuthAuthentication object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OAuthAuthentication$new(config, client_id = NULL, client_secret = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config}}{The client configuration}

\item{\code{client_id}}{The client ID}

\item{\code{client_secret}}{The client secret}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new OAuthAuthentication object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OAuthAuthentication-get_token_filename"></a>}}
\if{latex}{\out{\hypertarget{method-OAuthAuthentication-get_token_filename}{}}}
\subsection{Method \code{get_token_filename()}}{
Get the token filename
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OAuthAuthentication$get_token_filename()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The filename for storing this client's token
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OAuthAuthentication-load_saved_credentials"></a>}}
\if{latex}{\out{\hypertarget{method-OAuthAuthentication-load_saved_credentials}{}}}
\subsection{Method \code{load_saved_credentials()}}{
Load saved credentials from disk
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OAuthAuthentication$load_saved_credentials()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OAuthAuthentication-save_credentials"></a>}}
\if{latex}{\out{\hypertarget{method-OAuthAuthentication-save_credentials}{}}}
\subsection{Method \code{save_credentials()}}{
Save credentials to disk
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OAuthAuthentication$save_credentials()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OAuthAuthentication-refresh"></a>}}
\if{latex}{\out{\hypertarget{method-OAuthAuthentication-refresh}{}}}
\subsection{Method \code{refresh()}}{
Refresh the OAuth token
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OAuthAuthentication$refresh()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OAuthAuthentication-get_headers"></a>}}
\if{latex}{\out{\hypertarget{method-OAuthAuthentication-get_headers}{}}}
\subsection{Method \code{get_headers()}}{
Get the headers for HTTP requests
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OAuthAuthentication$get_headers()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A named character vector of headers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OAuthAuthentication-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OAuthAuthentication-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OAuthAuthentication$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
