% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-oauth_client.R
\name{OAuthClient}
\alias{OAuthClient}
\title{OAuthClient}
\description{
R6 class for OAuth client credentials
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the client}

\item{\code{description}}{The description of the client}

\item{\code{client_id}}{The client ID}

\item{\code{client_secret}}{The client secret}

\item{\code{enabled}}{Whether the client is enabled}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OAuthClient-new}{\code{OAuthClient$new()}}
\item \href{#method-OAuthClient-save}{\code{OAuthClient$save()}}
\item \href{#method-OAuthClient-to_list}{\code{OAuthClient$to_list()}}
\item \href{#method-OAuthClient-to_json}{\code{OAuthClient$to_json()}}
\item \href{#method-OAuthClient-clone}{\code{OAuthClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OAuthClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-OAuthClient-new}{}}}
\subsection{Method \code{new()}}{
Create a new OAuthClient object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OAuthClient$new(
  name = NULL,
  description = NULL,
  client_id = NULL,
  client_secret = NULL,
  enabled = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the client}

\item{\code{description}}{The description of the client}

\item{\code{client_id}}{The client ID}

\item{\code{client_secret}}{The client secret}

\item{\code{enabled}}{Whether the client is enabled}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new OAuthClient object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OAuthClient-save"></a>}}
\if{latex}{\out{\hypertarget{method-OAuthClient-save}{}}}
\subsection{Method \code{save()}}{
Save the OAuth client to disk
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OAuthClient$save(config = NULL, filename = API_SECRETS_FILE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config}}{The client config (optional)}

\item{\code{filename}}{The filename to save to (optional)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
self (invisible)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OAuthClient-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-OAuthClient-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert OAuthClient to a plain list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OAuthClient$to_list(include_secret = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{include_secret}}{Whether to include client_secret (default: FALSE for security)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list containing OAuth client fields
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OAuthClient-to_json"></a>}}
\if{latex}{\out{\hypertarget{method-OAuthClient-to_json}{}}}
\subsection{Method \code{to_json()}}{
Convert OAuthClient to JSON string
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OAuthClient$to_json(pretty = TRUE, include_secret = FALSE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{Whether to pretty-print the JSON (default: TRUE)}

\item{\code{include_secret}}{Whether to include client_secret (default: FALSE for security)}

\item{\code{...}}{Additional arguments passed to jsonlite::toJSON}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A JSON string representation of the OAuth client
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OAuthClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OAuthClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OAuthClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
