% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-project.R
\name{Project}
\alias{Project}
\title{Project}
\description{
R6 class representing a project
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{The project ID}

\item{\code{name}}{The project name (mapped from API 'title' field)}

\item{\code{description}}{The project description}

\item{\code{version}}{The project version}

\item{\code{api_version}}{The API version}

\item{\code{api_access}}{Whether API access is enabled}

\item{\code{variables}}{List of project variables}

\item{\code{exec_parameters}}{Execution parameters configuration}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Project-new}{\code{Project$new()}}
\item \href{#method-Project-print}{\code{Project$print()}}
\item \href{#method-Project-to_list}{\code{Project$to_list()}}
\item \href{#method-Project-to_json}{\code{Project$to_json()}}
\item \href{#method-Project-clone}{\code{Project$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Project-new"></a>}}
\if{latex}{\out{\hypertarget{method-Project-new}{}}}
\subsection{Method \code{new()}}{
Create a new Project object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$new(
  id = NULL,
  name = NULL,
  description = NULL,
  version = NULL,
  api_version = NULL,
  api_access = NULL,
  variables = NULL,
  exec_parameters = NULL,
  title = NULL,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The project ID}

\item{\code{name}}{The project name}

\item{\code{description}}{The project description}

\item{\code{version}}{The project version}

\item{\code{api_version}}{The API version}

\item{\code{api_access}}{Whether API access is enabled}

\item{\code{variables}}{List of project variables}

\item{\code{exec_parameters}}{Execution parameters}

\item{\code{title}}{Alternative name for 'name' field (API compatibility)}

\item{\code{...}}{Additional fields (ignored)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new Project object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Project-print"></a>}}
\if{latex}{\out{\hypertarget{method-Project-print}{}}}
\subsection{Method \code{print()}}{
Print method for Project
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments (ignored)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Project-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-Project-to_list}{}}}
\subsection{Method \code{to_list()}}{
Convert Project to a plain list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list containing all project fields
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Project-to_json"></a>}}
\if{latex}{\out{\hypertarget{method-Project-to_json}{}}}
\subsection{Method \code{to_json()}}{
Convert Project to JSON string
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$to_json(pretty = TRUE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pretty}}{Whether to pretty-print the JSON (default: TRUE)}

\item{\code{...}}{Additional arguments passed to jsonlite::toJSON}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A JSON string representation of the project
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Project-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Project-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
