% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{ServiceClientConfig}
\alias{ServiceClientConfig}
\title{ServiceClientConfig}
\description{
R6 class for Earth Blox API client configuration for service endpoints
}
\section{Super class}{
\code{ebx::ClientConfig} -> \code{ServiceClientConfig}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ServiceClientConfig-new}{\code{ServiceClientConfig$new()}}
\item \href{#method-ServiceClientConfig-clone}{\code{ServiceClientConfig$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ebx" data-topic="ClientConfig" data-id="get_api_base_url"><a href='../../ebx/html/ClientConfig.html#method-ClientConfig-get_api_base_url'><code>ebx::ClientConfig$get_api_base_url()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ebx" data-topic="ClientConfig" data-id="get_oauth_url"><a href='../../ebx/html/ClientConfig.html#method-ClientConfig-get_oauth_url'><code>ebx::ClientConfig$get_oauth_url()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ebx" data-topic="ClientConfig" data-id="get_persistence_driver"><a href='../../ebx/html/ClientConfig.html#method-ClientConfig-get_persistence_driver'><code>ebx::ClientConfig$get_persistence_driver()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ServiceClientConfig-new"></a>}}
\if{latex}{\out{\hypertarget{method-ServiceClientConfig-new}{}}}
\subsection{Method \code{new()}}{
Create a new ServiceClientConfig object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ServiceClientConfig$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new ServiceClientConfig object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ServiceClientConfig-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ServiceClientConfig-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ServiceClientConfig$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
