% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{set_client}
\alias{set_client}
\title{Set the current client}
\usage{
set_client(client, name = NULL)
}
\arguments{
\item{client}{The EbxClient object}

\item{name}{The client name (optional)}
}
\value{
The client (invisible)
}
\description{
Set a client as the current default
}
\keyword{internal}
